Move test files to src dir
