+++ jest.config.js
@@ -3,8 +3,6 @@ module.exports = {
   coverageDirectory: "coverage",
   coverageProvider: "v8",
   testMatch: [
-    "**/test/**/*.[jt]s?(x)",
-    "**/__tests__/**/*.[jt]s?(x)",
     "**/?(*.)+(spec|test).[jt]s?(x)"
   ],
   testTimeout: 2000,
